Some Basic Tests
This creates some basic unit tests for the read/write functionality. This uses
Boost::Test to execute, so it requires that the boost headers (>=1.59) be
there. Since this isn't common, the makefile will get them and depends on them
being in the boost_1_64_0 directory. Maybe someone who has experience with
makefiles can do this better.